update(figures): edits to multiple scripts for final figures
