+++ experiments/diffusion-experiments/classification-study.sh
@@ -1,0 +1,3 @@
+#!/bin/bash 
+
+python classification_user_study.py --ckpt-path "/home-local/Frederic/baselines/DiffusionModels/latent-guidance" --dataset-path "/home-local/Frederic/Datasets/FLCDataset/dataset-250k.tar"
