+++ experiments/diffusion-experiments/resolution_script.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#python resolution_experiment.py --weights MAE_SMALL_STED
+python resolution_experiment.py --weights MAE_SMALL_STED
 python resolution_experiment.py --weights MAE_SMALL_HPA
 python resolution_experiment.py --weights MAE_SMALL_SIM 
 python resolution_experiment.py --weights MAE_SMALL_JUMP
