+++ experiments/interpretability-experiments/segmentation_datasets/actin.py
@@ -9,6 +9,7 @@ from dataclasses import dataclass
 from skimage import io
 from torch.utils.data import Dataset
 from tqdm import tqdm
+from typing import Tuple, List
 from torchvision import transforms
 
 import sys
@@ -144,7 +145,7 @@ class HDF5Dataset(Dataset):
     def __len__(self):
         return len(self.samples)
     
-def get_dataset(cfg:dataclass, test_only:bool=False, **kwargs) -> tuple[Dataset, Dataset, Dataset]:
+def get_dataset(cfg:dataclass, test_only:bool=False, **kwargs) -> Tuple[Dataset, Dataset, Dataset]:
 
     # Updates the configuration inplace
     cfg.dataset_cfg = FActinConfiguration()
