+++ experiments/segmentation-experiments/alexa-script.sh
@@ -1,5 +1,5 @@
-python main.py --dataset synaptic-protein-segmentation --backbone mae-lightning-small --backbone-weights MAE_SMALL_JUMP --use-tensorboard --seed 42 
-#python main.py --dataset synaptic-protein-segmentation --backbone mae-lightning-small --backbone-weights MAE_SMALL_JUMP --use-tensorboard --seed 43 
-#python main.py --dataset synaptic-protein-segmentation --backbone mae-lightning-small --backbone-weights MAE_SMALL_JUMP --use-tensorboard --seed 44 
-#python main.py --dataset synaptic-protein-segmentation --backbone mae-lightning-small --backbone-weights MAE_SMALL_JUMP --use-tensorboard --seed 45 
-#python main.py --dataset synaptic-protein-segmentation --backbone mae-lightning-small --backbone-weights MAE_SMALL_JUMP --use-tensorboard --seed 46 
+python main.py --dataset factin --backbone mae-lightning-small --backbone-weights MAE_SMALL_STED --use-tensorboard --seed 42 
+python main.py --dataset factin --backbone mae-lightning-small --backbone-weights MAE_SMALL_STED --use-tensorboard --seed 43 
+python main.py --dataset factin --backbone mae-lightning-small --backbone-weights MAE_SMALL_STED --use-tensorboard --seed 44 
+python main.py --dataset factin --backbone mae-lightning-small --backbone-weights MAE_SMALL_STED --use-tensorboard --seed 45 
+python main.py --dataset factin --backbone mae-lightning-small --backbone-weights MAE_SMALL_STED --use-tensorboard --seed 46
