+++ experiments/segmentation-experiments/decoders/vit.py
@@ -157,7 +157,7 @@ class ViTDecoder(torch.nn.Module):
 class ViTSegmentationClassifier(torch.nn.Module):
     def __init__(self, backbone: torch.nn.Module, cfg: dataclass, global_pool: str = "patch") -> None:
         super().__init__()
-        self.backbone = backbone.vit 
+        self.backbone = backbone
         embed_dim = self.backbone.embed_dim 
         self.cfg = cfg 
         self.global_pool = global_pool
@@ -173,7 +173,6 @@ class ViTSegmentationClassifier(torch.nn.Module):
         )
 
     def forward_encoder(self, x: torch.Tensor) -> torch.Tensor:
-        print(x.shape)
         features = self.backbone.forward_features(x)
         if self.global_pool == "token":
             features = features[:, 0, :] # class token 
@@ -265,7 +264,9 @@ def get_decoder(backbone: torch.nn.Module, cfg: dataclass, **kwargs) -> torch.nn
     :returns : A `ViTDecoder` instance
     """
     if cfg.backbone in ["mae-lightning-tiny", "mae-lightning-small", "mae-lightning-base", "mae-lightning-large", "vit-tiny", "vit-small"]:
-        return ViTSegmentationClassifier(backbone=backbone.backbone, cfg=cfg)
+        if "MAE_SMALL_IMAGENET1K_V1" not in cfg.backbone_weights:
+            backbone = backbone.backbone.vit 
+        return ViTSegmentationClassifier(backbone=backbone, cfg=cfg)
         # extract_layers = [3, 6, 9 ,12]
         # return ViTDecoder(backbone.backbone, cfg, extract_layers=extract_layers, **kwargs)
     else:
